Refactor Schemas into separate native and realm-based implementations
