+++ realm/realm-library/src/main/java/io/realm/internal/RealmProxyMediator.java
@@ -31,6 +31,7 @@ import io.realm.RealmModel;
 import io.realm.RealmObject;
 import io.realm.RealmObjectSchema;
 import io.realm.RealmSchema;
+import io.realm.StandardRealmSchema;
 import io.realm.exceptions.RealmException;
