+++ realm/realm-library/src/main/java/io/realm/internal/SharedRealm.java
@@ -24,7 +24,7 @@ import java.util.List;
 import java.util.concurrent.CopyOnWriteArrayList;
 
 import io.realm.RealmConfiguration;
-import io.realm.RealmSchema;
+import io.realm.OsRealmSchema;
 import io.realm.internal.android.AndroidCapabilities;
 import io.realm.internal.android.AndroidRealmNotifier;
 
@@ -343,7 +343,7 @@ public final class SharedRealm implements Closeable, NativeObject {
      * Updates the underlying schema based on the schema description.
      * Calling this method must be done from inside a write transaction.
      */
-    public void updateSchema(RealmSchema schema, long version) {
+    public void updateSchema(OsRealmSchema schema, long version) {
         nativeUpdateSchema(nativePtr, schema.getNativePtr(), version);
     }
 
@@ -356,7 +356,7 @@ public final class SharedRealm implements Closeable, NativeObject {
         return nativeIsAutoRefresh(nativePtr);
     }
 
-    public boolean requiresMigration(RealmSchema schema) {
+    public boolean requiresMigration(OsRealmSchema schema) {
         return nativeRequiresMigration(nativePtr, schema.getNativePtr());
     }
